Add option of getting duration from a rss feed.
Add the setting feedURL which is the location of
an RSS feed. If there are <audio> tags which are
not set to preload, fetch the RSS feed and try to
find the duration from within the feed. Use that to
update the player times without having to make a
request to the media file.

This makes it possible to reduce the number of mp3
requests, which can save bandwidth and help with
stats.